ADMVEDU-422 add bottomNavigation in MainActivity.kt
