+++ app/build.gradle.kts
@@ -53,8 +53,10 @@ dependencies {
 
     implementation(Dependencies.fragment)
     implementation(Dependencies.ui)
+
     implementation(Dependencies.coroutines)
     implementation(Dependencies.runtime)
+
     implementation(Dependencies.core)
     implementation(Dependencies.appCompat)
     implementation(Dependencies.material)
