+++ app/src/main/AndroidManifest.xml
@@ -11,9 +11,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.ITransitionproject">
-        <activity android:name=".MainActivity" />
-        <activity android:name=".ui.login.LoginActivity"/>
-        <activity android:name=".ui.home.HomeActivity" />
+        <activity android:name=".ui.main.MainActivity" />
+        <activity android:name=".ui.login.LoginActivity" />
+        <activity android:name="com.example.itransition_project.ui.main.home.HomeActivity" />
         <activity
             android:name=".ui.splash.SplashScreenActivity"
             android:exported="true">
