--- app/src/main/java/com/example/itransition_project/MainActivity.kt
@@ -1,16 +1,0 @@
-package com.example.itransition_project
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.itransition_project.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-    lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-    }
-}
