+++ app/src/main/java/com/example/itransition_project/ui/main/MainActivity.kt
@@ -1,0 +1,29 @@
+package com.example.itransition_project.ui.main
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.NavigationUI
+import com.example.itransition_project.R
+import com.example.itransition_project.databinding.ActivityMainBinding
+import com.google.android.material.bottomnavigation.BottomNavigationView
+
+class MainActivity : AppCompatActivity() {
+
+    lateinit var binding: ActivityMainBinding
+    lateinit var navController: NavController
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottom_navigation_view)
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.mainContainerView) as NavHostFragment
+        navController = navHostFragment.navController
+
+        NavigationUI.setupWithNavController(bottomNavigationView, navController)
+
+    }
+}
