+++ app/src/main/java/com/example/itransition_project/ui/main/home/HomeActivity.kt
@@ -1,0 +1,14 @@
+package com.example.itransition_project.ui.main.home
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.itransition_project.R
+
+class HomeActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_home)
+
+    }
+}
