+++ app/src/main/java/com/example/itransition_project/ui/main/home/fragment/core/HomeFragment.kt
@@ -1,15 +1,17 @@
-package com.example.itransition_project.ui.home.fragment.core
+package com.example.itransition_project.ui.main.home.fragment.core
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.example.itransition_project.R
 import com.example.itransition_project.databinding.HomeFragmentBinding
-import com.example.itransition_project.databinding.LoginFragmentBinding
 
-class HomeFragment : Fragment() {
+class HomeFragment : Fragment(R.layout.home_fragment) {
+
     lateinit var binding: HomeFragmentBinding
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
