+++ app/src/main/java/com/example/itransition_project/ui/main/home/fragment/core/HomeViewModule.kt
@@ -1,0 +1,4 @@
+package com.example.itransition_project.ui.main.home.fragment.core
+
+class HomeViewModule {
+}
