+++ app/src/main/res/layout/activity_home.xml
@@ -6,19 +6,4 @@
     android:layout_height="match_parent">
 
 
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottom_navigation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:menu="@menu/bottom_navigation_menu"
-        tools:ignore="MissingConstraints" />
-
-    <FrameLayout
-        android:id="@+id/homeLayout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintBottom_toTopOf="@+id/bottom_navigation"/>
-
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
