+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/page_1"
+        android:id="@+id/homeFragment"
         android:enabled="true"
         android:title="@string/home"/>
     <item
-        android:id="@+id/page_2"
+        android:id="@+id/profileFragment"
         android:enabled="true"
         android:title="@string/profile"/>
 </menu>
