+++ app/src/main/res/navigation/nav_graph.xml
@@ -15,7 +15,7 @@
     </fragment>
     <activity
         android:id="@+id/mainActivity"
-        android:name="com.example.itransition_project.MainActivity"
+        android:name="com.example.itransition_project.ui.main.MainActivity"
         android:label="activity_main"
-        tools:layout="@layout/activity_main"/>
+        tools:layout="@layout/activity_main" />
 </navigation>
